Refactor get_elf_size
Use ssize_t to be able to add an error check (return value < 0).
Before, the function just called exit(), which is undesirable when
using from the library. Instead, the user of the library should be able
to treat the error gracefully.

Furthermore adds error checks everywhere the function is called.

+++ src/elf.h
@@ -16,6 +16,6 @@ e_shoff + ( e_shentsize * e_shnum ) =	126584
 #ifndef ELFSIZE_H
 #define ELFSIZE_H
 
-unsigned long get_elf_size(const char *fname);
+ssize_t get_elf_size(const char *fname);
 
 #endif /* ELFSIZE_H */
